% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\docType{data}
\name{adaptation}
\alias{adaptation}
\title{adaptation}
\format{
3 arguments across 18 observations
\describe{
\item{subject}{Subject ID}
\item{no}{No shock after stutter}
\item{following}{Shock following a stutter}
\item{during}{Shock during a stutter}
}
}
\usage{
adaptation
}
\description{
Table 7.10: Adaptation Scores for College-Age Stutterers
}
\keyword{datasets}
