% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aircon.R
\docType{data}
\name{aircon}
\alias{aircon}
\title{aircon}
\format{
1 argument across 12 observations
\describe{
\item{aircon}{List of Xi's}
}
}
\usage{
aircon
}
\description{
Table 11.2: Intervals in Hours between Failures of the Air-Conditioning System of Plane 8044
}
\keyword{datasets}
