% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albumin.R
\docType{data}
\name{albumin}
\alias{albumin}
\title{albumin}
\format{
3 arguments across 8 observations
\describe{
\item{patient}{Patient ID}
\item{xi}{Before}
\item{yi}{After}
}
}
\usage{
albumin
}
\description{
Table 3.18 - Intravascular Albumin Pool Before and After Prednisolone
}
\keyword{datasets}
