% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{alcohol}
\format{
1 arguments across 15 observations
\describe{
\item{control}{Alcohol consumption in the control group}
\item{sst}{Alcohol consumption in the social skills training group}
}
}
\usage{
alcohol
}
\description{
Table 4.2: Alcohol Intake for 1 Year (Centiliter of Pure Alcohol)
}
\keyword{datasets}
