% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcalcf.R
\docType{data}
\name{apcalcf}
\alias{apcalcf}
\title{apcalcf}
\format{
1 argument across 68 observations
\describe{
\item{apcalc}{List of Xi's}
}
}
\usage{
apcalcf
}
\description{
Table 12.1: Discrepancy Scores for 68 Female AP Calculus Students
}
\keyword{datasets}
