% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcalcm.R
\docType{data}
\name{apcalcm}
\alias{apcalcm}
\title{apcalcm}
\format{
1 argument across 82 observations
\describe{
\item{apcalc}{List of Xi's}
}
}
\usage{
apcalcm
}
\description{
Table 12.3: Discrepancy Scores for 82 Male AP Calculus Students
}
\keyword{datasets}
