% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\docType{data}
\name{assembly}
\alias{assembly}
\title{assembly}
\format{
4 arguments across 9 observations
\describe{
\item{worker}{Worker ID}
\item{a}{Assembly method type A}
\item{b}{Assembly method type B}
\item{c}{Assembly method type C}
\item{d}{Assembly method type D}
}
}
\usage{
assembly
}
\description{
Table 7.17: Assembly Times (min)
}
\keyword{datasets}
