% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beakclapping.R
\docType{data}
\name{beakclapping}
\alias{beakclapping}
\title{beakclapping}
\format{
4 arguments across 25 observations
\describe{
\item{embryo}{Embryo ID}
\item{xi}{The beak claps measured in the dark period}
\item{yi}{The beak claps measured in illumination}
\item{zi}{yi - xi}
}
}
\usage{
beakclapping
}
\description{
Table 3.5: Beak-Clapping Counts per Minute
}
\keyword{datasets}
