% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedmaking.R
\docType{data}
\name{bedmaking}
\alias{bedmaking}
\title{bedmaking}
\format{
2 arguments across 8 observations
\describe{
\item{patient}{Patient ID}
\item{zi}{cc's of oxygen}
}
}
\usage{
bedmaking
}
\description{
Table 3.15 - Net Oxygen Consumption (cc)
}
\keyword{datasets}
