% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleedingtime.R
\docType{data}
\name{bleedingtime}
\alias{bleedingtime}
\title{bleedingtime}
\format{
3 arguments across 6 observations
\describe{
\item{subject}{Subject ID}
\item{xi}{Before asprin}
\item{yi}{After asprin}
}
}
\usage{
bleedingtime
}
\description{
Table 3.7: Bleeding Time (in seconds)
}
\keyword{datasets}
