% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainweight.R
\docType{data}
\name{brainweight}
\alias{brainweight}
\title{brainweight}
\format{
3 arguments across 11 observations
\describe{
\item{subject}{Subject ID}
\item{weight}{Brain weight, grams}
\item{fiber}{Pyramidal large fiber count}
}
}
\usage{
brainweight
}
\description{
Table 8.4: Mean Brain Weights and Medullary Pyramid Large Fiber Counts for Cerebral Palsy Subjects
}
\keyword{datasets}
