% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catheter.R
\docType{data}
\name{catheter}
\alias{catheter}
\title{catheter}
\format{
4 arguments across 12 observations
\describe{
\item{child}{Child ID}
\item{height}{Height, inches}
\item{Weight}{Weight, pounds}
\item{length}{Length of heart catheter, centimeters}
}
}
\usage{
catheter
}
\description{
Table 9.10: Required Length of Heart Catheter as a Function of Height and Weight
}
\keyword{datasets}
