% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenosphereresin.R
\docType{data}
\name{cenosphereresin}
\alias{cenosphereresin}
\title{cenosphereresin}
\format{
3 arguments across 8 observations
\describe{
\item{specimen}{Specimen ID}
\item{pressure}{Pressue (psi)}
\item{density}{Density (grams/cubed-centimeter)}
}
}
\usage{
cenosphereresin
}
\description{
Table 9.2: The Effects of Hydrostatic Pressure on the Density of a Cenosphere-Resin Composite
}
\keyword{datasets}
