% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorioamnion.R
\docType{data}
\name{chorioamnion}
\alias{chorioamnion}
\title{chorioamnion}
\format{
1 arguments across 15 observations
\describe{
\item{term}{At term}
\item{later}{At 12-26 weeks gestational age}
}
}
\usage{
chorioamnion
}
\description{
Table 4.1: Tritiated Water Diffusion Across Human Chorioamnion
}
\keyword{datasets}
