% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciliarybeat.R
\docType{data}
\name{ciliarybeat}
\alias{ciliarybeat}
\title{ciliarybeat}
\format{
3 arguments across 15 observations
\describe{
\item{subject}{Subject ID}
\item{nasal}{Nasal brushing}
\item{biopsy}{Endobronchial forceps biopsy}
}
}
\usage{
ciliarybeat
}
\description{
Table 8.10: Relation between Ciliary Beat Frequency (CBF) Values Obtained through Nasal Brushing and Endobronchial Forceps Biopsy
}
\keyword{datasets}
