% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coatacidity.R
\docType{data}
\name{coatacidity}
\alias{coatacidity}
\title{coatacidity}
\format{
3 arguments across 18 observations
\describe{
\item{acidity}{Acidity}
\item{concentration}{Concentration of NaOH}
\item{type}{Type of coal}
}
}
\usage{
coatacidity
}
\description{
Table 7.23: Coat Acidity Value
}
\keyword{datasets}
