% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{concrete}
\format{
1 argument across 20 observations
\describe{
\item{concrete}{List of Xi's}
}
}
\usage{
concrete
}
\description{
Table 11.8: Pneumatic Pressures Required to Break Concrete Cubes
}
\keyword{datasets}
