% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronary.R
\docType{data}
\name{coronary}
\alias{coronary}
\title{coronary}
\format{
2 arguments across 10 observations
\describe{
\item{typea}{Subjects with Type A behavior}
\item{typeb}{Subjects with Type B behavior}
}
}
\usage{
coronary
}
\description{
Table 5.8: Peak Levels of Human Plasma Growth Hormone after Arginine Hydrochloride Infusion (Initial Test, nanograms per milliliter)
}
\keyword{datasets}
