% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotton.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{cotton}
\format{
1 arguments across 15 observations
\describe{
\item{p144}{Potash level of 144 pounds per acre (lb/acre)}
\item{p108}{Potash level of 108 pounds per acre (lb/acre)}
\item{p72}{Potash level of 72 pounds per acre (lb/acre)}
\item{p54}{Potash level of 54 pounds per acre (lb/acre)}
\item{p36}{Potash level of 36 pounds per acre (lb/acre)}
}
}
\usage{
cotton
}
\description{
Table 7.5: Strength Index of Cotton
}
\keyword{datasets}
