% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescribe.R
\docType{data}
\name{datadescribe}
\alias{datadescribe}
\title{Describe Dataset}
\format{
One argument
\describe{
\item{dataset}{An R data frame object}
}
}
\usage{
datadescribe
}
\description{
A function to get the quantitative description of a dataset
}
\keyword{datasets}
