% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{ddt}
\format{
2 arguments across 10 observations
\describe{
\item{worker}{Sample ID}
\item{zi}{Excretion density}
}
}
\usage{
ddt
}
\description{
Table 3.11 - 6-Beta-Hydroxycortisol Excretion (micrograms/24 hours)
}
\keyword{datasets}
