% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{describe}
\alias{describe}
\title{Describe Dataset}
\format{
One argument
\describe{
\item{dataset}{An R data frame object}
}
}
\usage{
describe(dataset)
}
\description{
A function to get the quantitative description of a dataset
}
