% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewoxidant.R
\docType{data}
\name{dewoxidant}
\alias{dewoxidant}
\title{dewoxidant}
\format{
2 arguments across 12 observations
\describe{
\item{sample}{Sample ID}
\item{zi}{Oxidant content, measured in ppm ozone}
}
}
\usage{
dewoxidant
}
\description{
Table 3.13 - Oxidant Content of Dew Water, Port Burwell, 1960
}
\keyword{datasets}
