% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs.R
\docType{data}
\name{dogs}
\alias{dogs}
\title{dogs}
\format{
3 arguments across 7 observations
\describe{
\item{dog}{Dog ID}
\item{xi}{The dog's before insulin level}
\item{yi}{The dog's after insulin level}
}
}
\usage{
dogs
}
\description{
Table 3.3: Blood Levels of Immunoreactive Insulin (microunits per milliliter)
}
\keyword{datasets}
