% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood.R
\docType{data}
\name{fastfood}
\alias{fastfood}
\title{fastfood}
\format{
1 argument across 34 observations
\describe{
\item{fastfood}{List of Xi's}
}
}
\usage{
fastfood
}
\description{
Table 11.15 (Abridged): Service Times at a Fast-Food Restaurant (Times Only)
}
\keyword{datasets}
