% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\docType{data}
\name{flux}
\alias{flux}
\title{flux}
\format{
3 arguments across 16 observations
\describe{
\item{core}{Core sample, i}
\item{time}{Time, xij (hours)}
\item{flux}{Ammonium flux, Yij (micromoles N/square-meter)}
}
}
\usage{
flux
}
\description{
Table 9.4: Coastal Sediment Ammonium Flux in Apalachicola Bay, Florida
}
\keyword{datasets}
