% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gizzard.R
\docType{data}
\name{gizzard}
\alias{gizzard}
\title{gizzard}
\format{
1 argument across 40 observations
\describe{
\item{i}{Length samples from Site I}
\item{ii}{Length samples from Site II}
\item{iii}{Length samples from Site III}
\item{iv}{Length samples from Site IV}
}
}
\usage{
gizzard
}
\description{
Table 6.3: Length of YOY Gizzard Shad from Kokosing Lake, Ohio, Sampled in Summer, 1984 (mm)
}
\keyword{datasets}
