% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glucoseimpedance.R
\docType{data}
\name{glucoseimpedance}
\alias{glucoseimpedance}
\title{glucoseimpedance}
\format{
3 arguments across 7 observations
\describe{
\item{subject}{Subject ID}
\item{response}{Weighted glucose response, X}
\item{impedance}{Glucose impedance, Y}
}
}
\usage{
glucoseimpedance
}
\description{
Table 8.12: Weighted Glucose Response and Glucose Impedance
}
\keyword{datasets}
