% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granulocytic.R
\docType{data}
\name{granulocytic}
\alias{granulocytic}
\title{granulocytic}
\format{
1 argument across 43 observations
\describe{
\item{granulocytic}{List of Xi's}
}
}
\usage{
granulocytic
}
\description{
Table 11.6: Ordered Survival Times (Days from Diagnosis)
}
\keyword{datasets}
