% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouse.R
\docType{data}
\name{grouse}
\alias{grouse}
\title{grouse}
\format{
2 arguments across 7 observations
\describe{
\item{grouse}{Grouse ID}
\item{zi}{Percentage time in active movement}
}
}
\usage{
grouse
}
\description{
Table 3.14 - Ruffed Grouse, Percentage Time in Active Movement
}
\keyword{datasets}
