% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinea.R
\docType{data}
\name{guinea}
\alias{guinea}
\title{guinea}
\format{
1 argument across 72 observations
\describe{
\item{guinea}{List of Xi's}
}
}
\usage{
guinea
}
\description{
Table 11.3: Ordered Survival Times in Days of Guinea Pigs under Regimen 4.3
}
\keyword{datasets}
