% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltondepression.R
\docType{data}
\name{hamiltondepression}
\alias{hamiltondepression}
\title{hamiltondepression}
\format{
3 arguments across 9 observations
\describe{
\item{patient}{The patient ID}
\item{xi}{Patient i's before value}
\item{yi}{Patient i's after value}
}
}
\usage{
hamiltondepression
}
\description{
Table 3.1: The Hamilton Depression Scale Factor IV Values
}
\keyword{datasets}
