% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebbwilliams.R
\docType{data}
\name{hebbwilliams}
\alias{hebbwilliams}
\title{hebbwilliams}
\format{
3 arguments across 36 observations
\describe{
\item{problem}{Problem ID}
\item{rats}{Error scores for rats}
\item{rabbits}{Error scores for rabbits}
\item{cats}{Error scores for cats}
}
}
\usage{
hebbwilliams
}
\description{
Table 7.9: Error Scores by Species
}
\keyword{datasets}
