% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemidiaphragms.R
\docType{data}
\name{hemidiaphragms}
\alias{hemidiaphragms}
\title{hemidiaphragms}
\format{
5 arguments across 12 observations
\describe{
\item{j}{Core sample, i}
\item{standardlog}{Standard X1j (log dose)}
\item{standardglycogen}{Standard Y1j (glycogen)}
\item{samplelog}{Sample I X2j (log dose)}
\item{sampleglycogen}{Sample I Y2j (glycogen)}
}
}
\usage{
hemidiaphragms
}
\description{
Table 9.6: Glycogen Content of Hemidiaphragms Measured by Optical Density in the Anthrone Test x 1000
}
\keyword{datasets}
