% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepatitis.R
\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{hepatitis}
\format{
4 argument across 24 observations
\describe{
\item{patient}{Patient ID}
\item{weeks}{Length of observation, weeks}
\item{status}{Status (1 = alive, 0 = dead)}
}
}
\usage{
hepatitis
}
\description{
Table 11.22: Severe Viral Hepatitis Study
}
\keyword{datasets}
