% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highschool.R
\docType{data}
\name{highschool}
\alias{highschool}
\title{highschool}
\format{
3 arguments across 18 observations
\describe{
\item{state}{State}
\item{dollars}{Dollars $ spent per Pupil}
\item{graduation}{Percentage Graduated}
}
}
\usage{
highschool
}
\description{
Table 8.6: Spending per High-School Senior and the Percentage of Those Seniors Who Graduated during the 1987-1988 School Year
}
\keyword{datasets}
