% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodgkins.R
\docType{data}
\name{hodgkins}
\alias{hodgkins}
\title{hodgkins}
\format{
3 argument across 49 observations
\describe{
\item{relapse}{Binary variable of whether the patient relapsed or not (where 1 = yes, 0 = no)}
\item{days}{Days healthy (if not relapsed) or days to relapse}
\item{radiation}{Whether radiation was to the affected node (affected) or it was total nodal radiation}
}
}
\usage{
hodgkins
}
\description{
Table 11.16: Relapse-Free Survival Times for Hodgkin's Disease Patients
}
\keyword{datasets}
