% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroxyproline.R
\docType{data}
\name{hydroxyproline}
\alias{hydroxyproline}
\title{hydroxyproline}
\format{
3 arguments across 7 observations
\describe{
\item{child}{Child ID}
\item{before}{The hydroxyproline density before growth hormone therapy}
\item{after}{The hydroxyproline density after growth hormone therapy}
}
}
\usage{
hydroxyproline
}
\description{
Table 3.4: Heat-Insoluble Hydroxyproline Micromoles per Gram of Dry Weight
}
\keyword{datasets}
