% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnotic.R
\docType{data}
\name{hypnotic}
\alias{hypnotic}
\title{hypnotic}
\format{
3 arguments across 6 observations
\describe{
\item{subject}{Subject ID}
\item{xi}{Before score}
\item{yi}{After score}
}
}
\usage{
hypnotic
}
\description{
Table 3.6: Average Scores on the Stanford Profile Scales of Hypnotic Susceptibility
}
\keyword{datasets}
