% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insecticide.R
\docType{data}
\name{insecticide}
\alias{insecticide}
\title{insecticide}
\format{
3 arguments across 12 observations
\describe{
\item{species}{Species of Drosophila fly}
\item{level}{Level of insecticide (ppm)}
\item{survived}{Number survived to adulthood}
}
}
\usage{
insecticide
}
\description{
Table 9.7: Numbers of Drosophila Flies (Three Different Species) That Survive to Adulthood after Exposure to Various Levels (ppm) of an Organic Phosphorus Insecticide
}
\keyword{datasets}
