% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironsup.R
\docType{data}
\name{ironsup}
\alias{ironsup}
\title{ironsup}
\format{
3 arguments across 109 observations
\describe{
\item{percentage}{Percentage of iron in sample}
\item{concentration}{Concentration in millimolars}
\item{form}{Form of iron (fe2, fe3)}
}
}
\usage{
ironsup
}
\description{
Table 7.21: Percentage of Iron Retained
}
\keyword{datasets}
