% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomers.R
\docType{data}
\name{isomers}
\alias{isomers}
\title{isomers}
\format{
5 arguments across 30 observations
\describe{
\item{run}{Experimental run ID}
\item{p250}{Conversion percentage at 250 PSI}
\item{p325}{Conversion percentage at 325 PSI}
\item{p400}{Conversion percentage at 400 PSI}
\item{p475}{Conversion percentage at 475 PSI}
\item{p550}{Conversion percentage at 550 PSI}
}
}
\usage{
isomers
}
\description{
Table 7.14: Percent Conversion of Methyl Glucoside to Monovinyl Isomers
}
\keyword{datasets}
