% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larvae.R
\docType{data}
\name{larvae}
\alias{larvae}
\title{larvae}
\format{
5 arguments across 14 observations
\describe{
\item{sample}{Sample ID}
\item{larvae}{Number of larvae of Chaoborus collected in a grab sample of the sediment from an area of approximately 225 square centimeters of lake bottom}
\item{depth}{Depth (meters) of the lake at the sampling point}
\item{brackishness}{Brackishness (conductivity) of the water at the lake bottom (recorded in mhos per decimeter)}
\item{oxygen}{Dissolved oxygen (milligrams per liter) in the water sampled from the lake bottom}
}
}
\usage{
larvae
}
\description{
Table 9.14: Number of Chaoborus Larvae and Water Quality of Samples
}
\keyword{datasets}
