% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadpoisoning.R
\docType{data}
\name{leadpoisoning}
\alias{leadpoisoning}
\title{leadpoisoning}
\format{
2 arguments across 15 observations
\describe{
\item{value}{Plasma glucose level of the goose}
\item{group}{Which goose was sampled (healthy, leadpoisoned)}
}
}
\usage{
leadpoisoning
}
\description{
Table 4.7: Plasma glucose values in healthy geese versus lead-poisoned geese
}
\keyword{datasets}
