% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukocyte.R
\docType{data}
\name{leukocyte}
\alias{leukocyte}
\title{leukocyte}
\format{
1 argument across 37 observations
\describe{
\item{normal}{GR sites per normal subject cell}
\item{hairy}{GR sites per hairy-cell anemia}
\item{lymphatic}{GR sites per chronic lymphatic leukemia cell}
\item{myelocytic}{GR sites per chronic myelocytic leukemia cell}
\item{acute}{GR sites per acute leukemia cell}
}
}
\usage{
leukocyte
}
\description{
Table 6.4: Number of Glucocorticoid Receptor (GR) Sites per Leukocyte Cell
}
\keyword{datasets}
