% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver.R
\docType{data}
\name{liver}
\alias{liver}
\title{liver}
\format{
6 arguments across 54 observations
\describe{
\item{patient}{Patient ID}
\item{time}{Survival time}
\item{clot}{A measure of the clotting potential of the patient's blood}
\item{prog}{A subjective index of the patient's prospect of recovery}
\item{enz}{A measure of a protein present in the body}
\item{liv}{A measure relating to white blood cell count}
}
}
\usage{
liver
}
\description{
Table 9.12: Survival Times of Liver Transplant Patients and Related Biological Measurements
}
\keyword{datasets}
