% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayfly.R
\docType{data}
\name{mayfly}
\alias{mayfly}
\title{mayfly}
\format{
2 arguments across 10 observations
\describe{
\item{mayfly}{Sample ID}
\item{zi}{Head width}
}
}
\usage{
mayfly
}
\description{
Table 3.10: Mayfly Head Width, Habitat A (Micrometer Divisions)
}
\keyword{datasets}
