% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic.R
\docType{data}
\name{metabolic}
\alias{metabolic}
\title{metabolic}
\format{
1 argument across 26 observations
\describe{
\item{X1}{Fasting metabolic rate between January-February}
\item{X2}{Fasting metabolic rate between March-April}
\item{X3}{Fasting metabolic rate between May-June}
\item{X4}{Fasting metabolic rate between July-August}
\item{X5}{Fasting metabolic rate between September-October}
\item{X6}{Fasting metabolic rate between November-December}
}
}
\usage{
metabolic
}
\description{
Table 6.8: Fasting Metabolic Rate (FMR) for White-Tailed Deer (kcal/kilogram/day)
}
\keyword{datasets}
