% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl.R
\docType{data}
\name{methyl}
\alias{methyl}
\title{methyl}
\format{
1 arguments across 10 observations
\describe{
\item{methyl}{List of integers with Xi data}
}
}
\usage{
methyl
}
\description{
Table 11.1: Calculation of Epsilon for the Methylmercury Data
}
\keyword{datasets}
