% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivational.R
\docType{data}
\name{motivational}
\alias{motivational}
\title{motivational}
\format{
1 argument across 18 observations
\describe{
\item{no}{Number of pieces processed by a subject with no information, control}
\item{rough}{Number of pieces processed by a subject in group B, rough information}
\item{accurate}{Number of pieces processed by a subject in group C, accurate information}
}
}
\usage{
motivational
}
\description{
Table 6.6: Number of Pieces Processed
}
\keyword{datasets}
