% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mucociliary.R
\docType{data}
\name{mucociliary}
\alias{mucociliary}
\title{mucociliary}
\format{
1 argument across 14 observations
\describe{
\item{normal}{Subjects with normal airways}
\item{obstructive}{Subjects with obstructive airways disease}
\item{asbestosis}{Subjects with asbestosis disease}
}
}
\usage{
mucociliary
}
\description{
Table 6.1: Half-Time of Mucociliary Clearance (hours)
}
\keyword{datasets}
