% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasturtiums.R
\docType{data}
\name{nasturtiums}
\alias{nasturtiums}
\title{nasturtiums}
\format{
7 arguments across 21 observations
\describe{
\item{day}{Day ID}
\item{a}{Chemical type A}
\item{b}{Chemical type B}
\item{c}{Chemical type C}
\item{d}{Chemical type D}
\item{e}{Chemical type E}
\item{f}{Chemical type F}
\item{g}{Chemical type G}
}
}
\usage{
nasturtiums
}
\description{
Table 7.12: Logarithm of Toxic Dosages
}
\keyword{datasets}
