% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niacin.R
\docType{data}
\name{niacin}
\alias{niacin}
\title{niacin}
\format{
3 arguments across 12 observations
\describe{
\item{lab}{Laboratory ID}
\item{none}{No niacin enrichment}
\item{four}{Four milligrams per 100g bran flakes of niacin enrichment}
\item{eight}{Eight milligrams per 100g bran flakes of niacin enrichment}
}
}
\usage{
niacin
}
\description{
Table 7.20: Amount of Niacin in Enriched Bran Flakes
}
\keyword{datasets}
