% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nursing.R
\docType{data}
\name{nursing}
\alias{nursing}
\title{nursing}
\format{
3 arguments across 18 observations
\describe{
\item{school}{School}
\item{decision}{Rank for faculty/dean decision-making agreement}
\item{satisfaction}{Rank for faculty satisfaction}
}
}
\usage{
nursing
}
\description{
Table 8.7: Rankings for Faculty/Dean Decision-Making Agreement and Faculty Satisfaction for Participating Schools of Nursing
}
\keyword{datasets}
