% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oakstands.R
\docType{data}
\name{oakstands}
\alias{oakstands}
\title{oakstands}
\format{
1 argument across 16 observations
\describe{
\item{X1}{Growing site index interval of 66-68}
\item{X2}{Growing site index interval of 69-71}
\item{X3}{Growing site index interval of 72-74}
\item{X4}{Growing site index interval of 75-77}
\item{X5}{Growing site index interval of 78-80}
}
}
\usage{
oakstands
}
\description{
Table 6.7: Average Basal Area Increment (BAI) Vlues for Oak Stands in Southeastern Ohio
}
\keyword{datasets}
