% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odors.R
\docType{data}
\name{odors}
\alias{odors}
\title{odors}
\format{
2 arguments across 15 observations
\describe{
\item{year}{Year}
\item{periods}{Number of odor periods}
}
}
\usage{
odors
}
\description{
Table 8.8: Annual Number of Odor Periods for Lake Michigan for the Period 1950-1964
}
\keyword{datasets}
