% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidant.R
\docType{data}
\name{oxidant}
\alias{oxidant}
\title{oxidant}
\format{
6 arguments across 30 observations
\describe{
\item{day}{Day ID}
\item{oxidant}{Maximum oxidant level}
\item{wind}{Wind speed}
\item{temperature}{Temperature (degrees F)}
\item{humidity}{Humidity}
\item{insolation}{Insolation (measure of the amount of sunlight)}
}
}
\usage{
oxidant
}
\description{
Table 9.13: Maximum Oxidant Level, Wind Speed, Temperature, Humidity, and Insolation for a 30-Day Summer Period in the Los Angeles Pollution Control District
}
\keyword{datasets}
