% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pine.R
\docType{data}
\name{pine}
\alias{pine}
\title{pine}
\format{
2 arguments across 36 observations
\describe{
\item{Length}{Interstitial length in centimeters}
\item{Type}{Type of pine tree}
}
}
\usage{
pine
}
\description{
Interstitial lengths of different species of pine tries
}
\keyword{datasets}
