% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasma.glucose.R
\docType{data}
\name{plasma.glucose}
\alias{plasma.glucose}
\title{plasma.glucose}
\format{
M arguments across N observations
\describe{
\item{zero}{Original plasma glucose level}
\item{one}{Plasma glucose level after 1 hour}
\item{four}{Plasma glucose level after 4 hours}
\item{twenty_four}{Plasma glucose level after 24 hours}
\item{ninety_six}{Plasma glucose level after 96 hours}
}
}
\usage{
plasma.glucose
}
\description{
Plasma glucose levels in geese after a certain number of hours
}
\keyword{datasets}
