% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeweed.R
\docType{data}
\name{pokeweed}
\alias{pokeweed}
\title{pokeweed}
\format{
1 argument across 10 observations
\describe{
\item{kentucky}{Kentucky pokeweed}
\item{florida}{Florida pokeweed}
}
}
\usage{
pokeweed
}
\description{
Table 5.2 - Average Dry Feces Weight (mg)
}
\keyword{datasets}
