% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{pollution}
\format{
4 arguments across 8 observations
\describe{
\item{year}{The year measured}
\item{odorperiods}{The number of odor periods measured in that year}
}
}
\usage{
pollution
}
\description{
Annual number of odor periods for Lake Michigan for the period 1950-1964
}
\keyword{datasets}
