% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prednisone.R
\docType{data}
\name{prednisone}
\alias{prednisone}
\title{prednisone}
\format{
1 arguments across 16 observations
\describe{
\item{with}{Mothers given prednisone}
\item{without}{Mothers not given prednisone}
}
}
\usage{
prednisone
}
\description{
Table 5.4: Platelet Counts of Newbonr Infants (per Millimeter-cubed)
}
\keyword{datasets}
