% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proline.R
\docType{data}
\name{proline}
\alias{proline}
\title{proline}
\format{
3 arguments across 7 observations
\describe{
\item{patient}{Patient ID}
\item{total}{Total collage, Xi, (mg/g dry weight of liver)}
\item{free}{Free proline, Yi, (micromoles/dry weight of liver)}
}
}
\usage{
proline
}
\description{
Table 8.9: Free Proline and Total Collagen Contents of Cirrhotic Patients
}
\keyword{datasets}
