% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychotherapy.R
\docType{data}
\name{psychoactive}
\alias{psychoactive}
\title{psychoactive}
\format{
4 arguments across 8 observations
\describe{
\item{control}{The control group}
\item{TR}{Reading}
\item{VTP}{Videotape}
\item{RII}{Group}
}
}
\usage{
psychoactive
}
\description{
Raw scores indicating the degree of psychotherapeutic attraction for each experimental condition
}
\keyword{datasets}
