% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychotic.R
\docType{data}
\name{psychotic}
\alias{psychotic}
\title{psychotic}
\format{
5 arguments across 14 observations
\describe{
\item{subject}{Subject ID}
\item{pre}{Preexercise}
\item{post19}{19 hours postexercise}
\item{post42}{42 hours postexercise}
\item{peak}{Peak-psychotic period}
}
}
\usage{
psychotic
}
\description{
Table 7.3: Effect of Isometric Exercise on Serum Creating Phosphokinase (CPK) Activity (mU/liter) in Psychotic Patients
}
\keyword{datasets}
