% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratskin.R
\docType{data}
\name{ratskin}
\alias{ratskin}
\title{ratskin}
\format{
7 arguments across 21 observations
\describe{
\item{rat}{Experimental run ID}
\item{a}{Chemical type A}
\item{b}{Chemical type B}
\item{c}{Chemical type C}
\item{d}{Chemical type D}
\item{e}{Chemical type E}
\item{f}{Chemical type F}
\item{g}{Chemical type G}
}
}
\usage{
ratskin
}
\description{
Table 7.13: Reactions of Male Rats to Chemical Substances
}
\keyword{datasets}
