% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundingfirst.R
\docType{data}
\name{roundingfirst}
\alias{roundingfirst}
\title{roundingfirst}
\format{
Seven arguments across twenty-two observations
\describe{
\item{Players}{The player}
\item{roundout}{Player's round out run time in seconds}
\item{roundout.rank}{Player's round out run time rank (by player)}
\item{narrowangle}{Player's narrow angle run time in seconds}
\item{narrowangle.rank}{Player's narrow angle run time rank (by player)}
\item{wideangle}{Player's wide angle run time in seconds}
\item{wideangle.rank}{Player's wide angle run time rank (by player)}
}
}
\usage{
roundingfirst
}
\description{
Table 7.1: Rounding First-Base Times
}
\keyword{datasets}
