% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salaries.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{salaries}
\format{
3 arguments across 12 observations
\describe{
\item{pair}{Pair ID}
\item{private}{The equivalent private sector salary}
\item{government}{The equivalent government job salary}
}
}
\usage{
salaries
}
\description{
Table 3.2: Annual Salaries
}
\keyword{datasets}
