% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salivation.R
\docType{data}
\name{salivation}
\alias{salivation}
\title{salivation}
\format{
2 arguments across 10 observations
\describe{
\item{feedback}{Salivation rate for group with feedback}
\item{nofeedback}{Salivation rate for group with no feedback}
}
}
\usage{
salivation
}
\description{
Table 5.7: Mean Drop Differences
}
\keyword{datasets}
