% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledataset.R
\docType{data}
\name{sample}
\alias{sample}
\title{sample}
\format{
M arguments across N observations
\describe{
\item{Variable}{Variable desciption}
\item{Variable}{Variable description}
\item{AnotherVariable}{Add more of these rows as needed}
}
}
\usage{
sample
}
\description{
Table K.L: Sample dataset description is written here.
}
\keyword{datasets}
