% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serum.R
\docType{data}
\name{serum}
\alias{serum}
\title{serum}
\format{
3 arguments across 60 observations
\describe{
\item{dosage}{Serum level of luteinizing hormone}
\item{regime}{Light regime (constant or 14 h light/10 h dark)}
\item{hormone}{Dosage of luteinizing release factor (LRF)}
}
}
\usage{
serum
}
\description{
Table 7.24: Serum Level of LH (in Nanograms per Milliliter of Serum)
}
\keyword{datasets}
