% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serumiron.R
\docType{data}
\name{serumiron}
\alias{serumiron}
\title{serumiron}
\format{
1 arguments across 40 observations
\describe{
\item{ramsay}{Serum iron detected using Ramsay method}
\item{jungparekh}{Serum iron detected using Jung-Parekh method}
}
}
\usage{
serumiron
}
\description{
Table 5.1: Serum Iron (micrograms per 100 milliliters) Determination Using Hyland Control Sera
}
\keyword{datasets}
