% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlingvelocity.R
\docType{data}
\name{settlingvelocity}
\alias{settlingvelocity}
\title{settlingvelocity}
\format{
2 arguments across 7 observations
\describe{
\item{sample}{Sample ID}
\item{zi}{Settling velocity, measured in centimeters/second}
}
}
\usage{
settlingvelocity
}
\description{
Table 3.12 - Settling Velocities of Sand Samples at 22-degrees Celcius
}
\keyword{datasets}
