% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokers.R
\docType{data}
\name{smokers}
\alias{smokers}
\title{smokers}
\format{
2 arguments across 22 observations
\describe{
\item{sputumhistaminelevel}{Level in microgram/gram}
\item{allergic}{Groups of allergic/nonallergic}
}
}
\usage{
smokers
}
\description{
Sputum histamine levels (microgram/gram of dry weight sputum)
}
\keyword{datasets}
