% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowy.R
\docType{data}
\name{snowy}
\alias{snowy}
\title{snowy}
\format{
2 arguments across 5 observations
\describe{
\item{years}{Years seeded, xi}
\item{double}{Double ratio, Yi}
}
}
\usage{
snowy
}
\description{
Table 9.1: Double Ratio for 5 Years in the Snowy Mountains of Australia
}
\keyword{datasets}
