% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodiumion.R
\docType{data}
\name{sodiumion}
\alias{sodiumion}
\title{sodiumion}
\format{
2 arguments across 10 observations
\describe{
\item{plasma}{Plasma sodium ion determination}
\item{erythrocyte}{Erythrocyte sodium ion determination}
}
}
\usage{
sodiumion
}
\description{
Table 5.3: Sodium Ion Content (mequiv/liter)
}
\keyword{datasets}
