% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrelmonkey.R
\docType{data}
\name{squirrelmonkey}
\alias{squirrelmonkey}
\title{squirrelmonkey}
\format{
3 arguments across 9 observations
\describe{
\item{monkey}{Monkey ID}
\item{weight}{Body weight, g}
\item{surface}{Total surface area, cubed-centimeters}
}
}
\usage{
squirrelmonkey
}
\description{
Table 9.3: Body Weight and Total Surface Area of Squirrel Monkeys
}
\keyword{datasets}
