% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stainless.R
\docType{data}
\name{stainless}
\alias{stainless}
\title{stainless}
\format{
2 arguments across 12 observations
\describe{
\item{sample}{Sample ID}
\item{percent}{Percentage of chromium}
}
}
\usage{
stainless
}
\description{
Table 3.9: Percentage of Chromium in the Stainless Steel Samples
}
\keyword{datasets}
