% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockreturn.R
\docType{data}
\name{stockreturn}
\alias{stockreturn}
\title{stockreturn}
\format{
3 arguments across 32 observations
\describe{
\item{company}{Company ID}
\item{return}{Mean rate percent of return, 1956-1969}
\item{value}{Value of common stock portfolio, December 31, 1969 (millions of dollars)}
}
}
\usage{
stockreturn
}
\description{
Table 8.11: Mean Rate of Return of Common Stock Portfolios over the Period 1956-1969 and the 1969 Value of Each Equity Portfolio for 32 Life Insurance Companies
}
\keyword{datasets}
