% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuttering.R
\docType{data}
\name{stuttering}
\alias{stuttering}
\title{stuttering}
\format{
4 arguments across 12 observations
\describe{
\item{subject}{Subject ID}
\item{r}{Subject spoke with a regular (rhythmic) metronome set at 120 ticks per minute}
\item{a}{Subject spoke with an arrhythmic metronome with intervals between 0.3 and 0.7 s but with an average of 120 ticks per minute}
\item{n}{Subject spoke unaided by a metronome}
}
}
\usage{
stuttering
}
\description{
Table 7.6: Influence of Rhythmicity of Metronome on Speech Fluency
}
\keyword{datasets}
