% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{survival}
\alias{survival}
\title{survival}
\format{
1 argument across 72 observations
\describe{
\item{survival}{List of Xi's}
}
}
\usage{
survival
}
\description{
Table 11.11: Ordered Survival Times in Days of Guinea Pigs under Regimen 6.6
}
\keyword{datasets}
