% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapeworms.R
\docType{data}
\name{tapeworms}
\alias{tapeworms}
\title{tapeworms}
\format{
3 arguments across 18 observations
\describe{
\item{dog}{Dog ID}
\item{cysticerci}{Weight of cysticerci fed to dogs}
\item{worms}{Weight of worms recovered at 20 days}
}
}
\usage{
tapeworms
}
\description{
Table 8.3: Relation Between Weight of the Cysticerci of Taenia hydatigena Fed to Dogs and Weight of Worms Recovered at 20 Days
}
\keyword{datasets}
