\name{toc}
\alias{table_of_contents}
\title{Table of Contents}
\usage{
help(toc)
}
\arguments{
\item{smokers}{Table 4.3 - Sputum Histamine Levels (micro-grams/gram Dry Weight Sputum)}
\item{violence}{Table 4.4 - Seconds Spent in Room after Witnessing Violence}
}
\value{
None
}
\description{
A guide to the datasets in nsm3data. Datasets are listed under the Arguments section by the order they appear in the Nonparametric Statistical Methods textbook, 3ed. The reference name in the \code{nsm3data} package is given, along with the table number and the descriptor of the dataset.
}
