% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triglyceride.R
\docType{data}
\name{triglyceride}
\alias{triglyceride}
\title{triglyceride}
\format{
9 arguments across 13 observations
\describe{
\item{patient}{Patient ID}
\item{total}{Total triglyceride level, mmol/liter}
\item{sex}{Sex of the patient (coded as female = 0, male = 1)}
\item{obese}{Whether patient is obese (coded as no = 0, yes = 1)}
\item{clylomicrons}{Chylo-microns}
\item{vldl}{Very low density lipoprotein}
\item{ldl}{Low density lipoprotein}
\item{hdl}{High density lipoprotein}
\item{age}{Age of the patient}
}
}
\usage{
triglyceride
}
\description{
Table 9.11: Blood Plasma Measurements Related to Total Triglyceride Level
}
\keyword{datasets}
