% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubercle.R
\docType{data}
\name{tubercle}
\alias{tubercle}
\title{tubercle}
\format{
1 argument across 72 observations
\describe{
\item{tubercle}{List of Xi's}
}
}
\usage{
tubercle
}
\description{
Table 11.9: Ordered Survival Times in Days of Guinea Pigs under Regimen 5.5
}
\keyword{datasets}
