% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuna.R
\docType{data}
\name{tuna}
\alias{tuna}
\title{tuna}
\format{
2arguments across 9 observations
\describe{
\item{lot}{Lot number for the can}
\item{hunter}{Hunter's L value}
\item{panel}{Panel score for the tuna}
}
}
\usage{
tuna
}
\description{
Table 8.1: Hunter L Values and Consumer Panel Scores for Nine Lots of Canned Tuna
}
\keyword{datasets}
