% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twins.R
\docType{data}
\name{twins}
\alias{twins}
\title{twins}
\format{
2 arguments across 13 observations
\describe{
\item{TwinX}{Exam score for twin X}
\item{TwinY}{Exam score for twin Y}
}
}
\usage{
twins
}
\description{
Table 8.5: Paired exam data of 13 different twin pairs
}
\keyword{datasets}
