% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violence.R
\docType{data}
\name{violence}
\alias{violence}
\title{violence}
\format{
2 arguments across 42 observations
\describe{
\item{seconds}{Number of seconds child remained in room}
\item{media}{Which program the child watched (olympics, karatekid)}
}
}
\usage{
violence
}
\description{
Table 4.4: Seconds spent in the room after witnessing violence
}
\keyword{datasets}
