% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welds.R
\docType{data}
\name{welds}
\alias{welds}
\title{welds}
\format{
3 arguments across 30 observations
\describe{
\item{strength}{Weld strength}
\item{cycle}{Weld cycle time}
\item{gage}{Gage bar setting}
}
}
\usage{
welds
}
\description{
Table 7.22: Strength of Weld
}
\keyword{datasets}
