% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescribe.R
\docType{data}
\name{datadescribe}
\alias{datadescribe}
\title{Describe Dataset}
\format{
An object of class \code{function} of length 1.
}
\usage{
datadescribe
}
\arguments{
\item{dataset}{A dataframe to describe}
}
\value{
nothing
}
\description{
A function to get the quantitative description of a dataset
}
\keyword{datasets}
