% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnum.R
\name{tabnum}
\alias{tabnum}
\title{Table numbers}
\format{
No arguments
\describe{
\item{null}{null}
}
}
\usage{
tabnum()
}
\description{
A function to alias all datasets with their corresponding table numbers. All table numbers are prefaced with a `t` (for example, instead of `4.5`, use `t4.5`). Test tables are prefaced with a `s`
}
