% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Table of Contents}
\usage{
help(toc)
}
\arguments{
\item{hamiltondepression}{Table 3.1 - The Hamilton Depression Scale Factor IV Values}

\item{salaries}{Table 3.2 - Annual Salaries}

\item{dogs}{Table 3.3 - Blood Levels of Immunoreactive Insuline (microunits per milliliter)}

\item{hydroxyproline}{Table 3.4 - Heat-Insoluble Hydroxyproline Micromoles per Gram of Dry Weight}

\item{beakclapping}{Table 3.5 - Beak-Clapping Counts per Minute}

\item{hypnotic}{Table 3.6 - Average Scores on the Stanford Profile Scales of Hypnotic Susceptibility}

\item{bleedingtime}{Table 3.7 - Bleeding Time (in seconds)}

\item{stainless}{Table 3.9 - Percentage of Chromium in the Stainless Steel Samples}

\item{mayfly}{Table 3.10 - Mayfly Head Width, Habitat A (Micrometer Divisions)}

\item{ddt}{Table 3.11 - 6-Beta-Hydroxycortisol Excretion (micrograms/24 hours)}

\item{settlingvelocity}{Table 3.12 - Settling Velocities of Sand Samples at 22-degrees Celcius}

\item{dewoxidant}{Table 3.13 - Oxidant Content of Dew Water, Port Burwell, 1960}

\item{grouse}{Table 3.14 - Ruffed Grouse, Percentage Time in Active Movement}

\item{bedmaking}{Table 3.15 - Net Oxygen Consumption (cc)}

\item{albumin}{Table 3.18 - Intravascular Albumin Pool Before and After Prednisolone}

\item{chorioamnion}{Table 4.1 - Tritiated Water Diffusion Across Human Chorioamnion}

\item{alcohol}{Table 4.2: Alcohol Intake for 1 Year (Centiliter of Pure Alcohol)}

\item{smokers}{Table 4.3 - Sputum Histamine Levels (micro-grams/gram Dry Weight Sputum)}

\item{violence}{Table 4.4 - Seconds Spent in Room after Witnessing Violence}

\item{leadpoisoning}{Table 4.7 - Plasma Glucose Levels}

\item{serumiron}{Table 5.1: Serum Iron (micrograms per 100 milliliters) Determination Using Hyland Control Sera}

\item{pokeweed}{Table 5.2 - Average Dry Feces Weight (mg)}

\item{prednisone}{Table 5.4: Platelet Counts of Newbonr Infants (per Millimeter-cubed)}
}
\description{
\code{help(toc)} displays the table of contents for datasets.
}
\details{
A guide to the datasets in nsm3data. Datasets are listed under the Arguments section by the order they appear in the Nonparametric Statistical Methods textbook, 3ed. The reference name in the \code{nsm3data} package is given, along with the table number and the descriptor of the dataset.
}
