% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{toc}
\usage{
help(toc)
}
\arguments{
\item{hamiltondepression}{Table 3.1 - The Hamilton Depression Scale Factor IV Values}

\item{salaries}{Table 3.2 - Annual Salaries}

\item{dogs}{Table 3.3 - Blood Levels of Immunoreactive Insuline (microunits per milliliter)}

\item{hydroxyproline}{Table 3.4 - Heat-Insoluble Hydroxyproline Micromoles per Gram of Dry Weight}

\item{beakclapping}{Table 3.5 - Beak-Clapping Counts per Minute}

\item{hypnotic}{Table 3.6 - Average Scores on the Stanford Profile Scales of Hypnotic Susceptibility}

\item{bleedingtime}{Table 3.7 - Bleeding Time (in seconds)}

\item{stainless}{Table 3.9 - Percentage of Chromium in the Stainless Steel Samples}

\item{mayfly}{Table 3.10 - Mayfly Head Width, Habitat A (Micrometer Divisions)}

\item{ddt}{Table 3.11 - 6-Beta-Hydroxycortisol Excretion (micrograms/24 hours)}

\item{settlingvelocity}{Table 3.12 - Settling Velocities of Sand Samples at 22-degrees Celcius}

\item{dewoxidant}{Table 3.13 - Oxidant Content of Dew Water, Port Burwell, 1960}

\item{grouse}{Table 3.14 - Ruffed Grouse, Percentage Time in Active Movement}

\item{bedmaking}{Table 3.15 - Net Oxygen Consumption (cc)}

\item{albumin}{Table 3.18 - Intravascular Albumin Pool Before and After Prednisolone}

\item{chorioamnion}{Table 4.1 - Tritiated Water Diffusion Across Human Chorioamnion}

\item{alcohol}{Table 4.2: Alcohol Intake for 1 Year (Centiliter of Pure Alcohol)}

\item{smokers}{Table 4.3 - Sputum Histamine Levels (micro-grams/gram Dry Weight Sputum)}

\item{violence}{Table 4.4 - Seconds Spent in Room after Witnessing Violence}

\item{leadpoisoning}{Table 4.7 - Plasma Glucose Levels}

\item{serumiron}{Table 5.1: Serum Iron (micrograms per 100 milliliters) Determination Using Hyland Control Sera}

\item{pokeweed}{Table 5.2 - Average Dry Feces Weight (mg)}

\item{sodiumion}{Table 5.3: Sodium Ion Content (mequiv/liter)}

\item{prednisone}{Table 5.4: Platelet Counts of Newbonr Infants (per Millimeter-cubed)}

\item{salivation}{Table 5.7: Mean Drop Differences}

\item{coronary}{Table 5.8: Peak Levels of Human Plasma Growth Hormone after Arginine Hydrochloride Infusion (Initial Test, nanograms per milliliter)}

\item{mucociliary}{Table 6.1: Half-Time of Mucociliary Clearance (hours)}

\item{psychotherapy}{Table 6.2: Raw scores indicating the degree of psychotherapeutic attraction for each experimental condition}

\item{gizzard}{Table 6.3: Length of YOY Gizzard Shad from Kokosing Lake, Ohio, Sampled in Summer, 1984 (mm)}

\item{leukocyte}{Table 6.4: Number of Glucocorticoid Receptor (GR) Sites per Leukocyte Cell}

\item{pine}{Table 6.5: Mean Interstitial Lengths (mm)}

\item{motivational}{Table 6.6: Number of Pieces Processed}

\item{oakstands}{Table 6.7: Average Basal Area Increment (BAI) Vlues for Oak Stands in Southeastern Ohio}

\item{metabolic}{Table 6.8: Fasting Metabolic Rate (FMR) for White-Tailed Deer (kcal/kilogram/day)}

\item{plasma.glucose}{Table 6.9: Plasma glucose levels in muskellungefish after a certain number of hours}

\item{acidred}{Table 6.10: Number of Revertant Colonies of Salmonella Bacteria of Strain TA98 under Exposure to Various Doses of Acid Red 114, with Hamster Liver Activation}

\item{roundingfirst}{Table 7.1: Rounding First-Base Times}

\item{ozone}{Table 7.2: Effect of Experimental Ozone Exposures on Airway Resistance (centimeters of a liter of water/second)}

\item{psychotic}{Table 7.3: Effect of Isometric Exercise on Serum Creating Phosphokinase (CPK) Activity (mU/liter) in Psychotic Patients}

\item{syllables}{Table 7.4: Percentage Consonants Correctly Identified under Each of the Conditions: (A) Audition, (L) Lip Reading, (AL) Audition and Lip Reading, (C) Cued Speech, (AC) Audition and Cued Speech, (LC) Lip Reading and Cued Speech, and (ALC) Audition, Lip Reading, and Cued Speech}

\item{cotton}{Table 7.5: Strength Index of Cotton}

\item{stuttering}{Table 7.6: Influence of Rhythmicity of Metronome on Speech Fluency}

\item{shelterbelts}{Table 7.7: Percent reduction in average wind speed at Dambutta, 1980/81 by shelterbelt trees}

\item{soiltemp}{Table 7.8: Maximum Soil Temperature (degrees Centigrade) at 5-cm Depth at Dambatta, 1980/81}

\item{hebbwilliams}{Table 7.9: Error Scores by Species}

\item{adaptation}{Table 7.10: Adaptation Scores for College-Age Stutterers}

\item{nasturtiums}{Table 7.12: Logarithm of Toxic Dosages}

\item{ratskin}{Table 7.13: Reactions of Male Rats to Chemical Substances}

\item{isomers}{Table 7.14: Percent Conversion of Methyl Glucoside to Monovinyl Isomers}

\item{shelterbeltsnov}{Table 7.16: Percent reduction in average wind speed at Dambutta, 1980/81 by shelterbelt trees, November included}

\item{assembly}{Table 7.17: Assembly Times (min)}

\item{niacin}{Table 7.20: Amount of Niacin in Enriched Bran Flakes}

\item{ironsup}{Table 7.21: Percentage of Iron Retained}

\item{welds}{Table 7.22: Strength of Weld}

\item{coatacidity}{Table 7.23: Coat Acidity Value}

\item{serum}{Table 7.24: Serum Level of LH (in Nanograms per Milliliter of Serum)}

\item{tremors}{Table 7.26: Forearm Tremor Frequency (Hz) as a Function of Weight Applied at the Wrist}

\item{tuna}{Table 8.1: Hunter L Values and Consumer Panel Scores for Nine Lots of Canned Tuna}

\item{tapeworms}{Table 8.3: Relation Between Weight of the Cysticerci of Taenia hydatigena Fed to Dogs and Weight of Worms Recovered at 20 Days}

\item{brainweight}{Table 8.4: Mean Brain Weights and Medullary Pyramid Large Fiber Counts for Cerebral Palsy Subjects}

\item{twins}{Table 8.5: Paired exam data of 13 different twin pairs}

\item{highschool}{Table 8.6: Spending per High-School Senior and the Percentage of Those Seniors Who Graduated during the 1987-1988 School Year}

\item{nursing}{Table 8.7: Rankings for Faculty/Dean Decision-Making Agreement and Faculty Satisfaction for Participating Schools of Nursing}

\item{odors}{Table 8.8: Annual Number of Odor Periods for Lake Michigan for the Period 1950-1964}

\item{prolineTable}{8.9: Free Proline and Total Collagen Contents of Cirrhotic Patients}

\item{ciliarybeat}{Table 8.10: Relation between Ciliary Beat Frequency (CBF) Values Obtained through Nasal Brushing and Endobronchial Forceps Biopsy}

\item{stockreturn}{Table 8.11: Mean Rate of Return of Common Stock Portfolios over the Period 1956-1969 and the 1969 Value of Each Equity Portfolio for 32 Life Insurance Companies}

\item{glucoseimpedance}{Table 8.12: Weighted Glucose Response and Glucose Impedance}

\item{snowy}{Table 9.1: Double Ratio for 5 Years in the Snowy Mountains of Australia}

\item{cenosphereresin}{Table 9.2: The Effects of Hydrostatic Pressure on the Density of a Cenosphere-Resin Composite}

\item{squirrelmonkey}{Table 9.3: Body Weight and Total Surface Area of Squirrel Monkeys}

\item{flux}{Table 9.4: Coastal Sediment Ammonium Flux in Apalachicola Bay, Florida}

\item{cloudseeding}{Table 9.5: Precipitation Amounts RI and Circulation Index M for Seeded and Control Units}

\item{hemidiaphragms}{Table 9.6: Glycogen Content of Hemidiaphragms Measured by Optical Density in the Anthrone Test x 1000}

\item{insecticide}{Table 9.7: Numbers of Drosophila Flies (Three Different Species) That Survive to Adulthood after Exposure to Various Levels (ppm) of an Organic Phosphorus Insecticide}

\item{catheter}{Table 9.10: Required Length of Heart Catheter as a Function of Height and Weight}

\item{triglyceride}{Table 9.11: Blood Plasma Measurements Related to Total Triglyceride Level}

\item{liver}{Table 9.12: Survival Times of Liver Transplant Patients and Related Biological Measurements}

\item{oxidant}{Table 9.13: Maximum Oxidant Level, Wind Speed, Temperature, Humidity, and Insolation for a 30-Day Summer Period in the Los Angeles Pollution Control District}

\item{larvae}{Table 9.14: Number of Chaoborus Larvae and Water Quality of Samples}
}
\description{
\code{help(toc)} displays the table of contents for datasets.
}
\details{
A guide to the datasets in nsm3data. Datasets are listed under the Arguments section by the order they appear in the Nonparametric Statistical Methods textbook, 3ed. The reference name in the \code{nsm3data} package is given, along with the table number and the descriptor of the dataset.
}
